requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
torch>=1.10.0
streamlit>=1.10.0
pandas>=1.3.0
Pillow>=8.0.0
google-generativeai>=0.3.0 